user router dsipatch model method to model/user
