+++ server/controllers/users.go
@@ -7,32 +7,19 @@ import (
 	"github.com/shifujito/chat_go/server/model"
 )
 
-type User struct {
-	Id   uint   `json:"id"`
-	Name string `json:"name"`
-}
-
 func UserHandler(w http.ResponseWriter, r *http.Request) {
+	corsSetup(w)
 	if r.Method == "GET" {
-		findAllUser(w)
+		getUser(w)
 	}
 }
 
-func findAllUser(w http.ResponseWriter) {
+func getUser(w http.ResponseWriter) {
 	// jsonをかえす。
-	modelUsers := []model.User{}
-	users := []User{}
-	db := model.DbConnect()
-	db.Find(&modelUsers)
-	for _, val := range modelUsers {
-		user := User{Id: val.ID, Name: val.Name}
-		users = append(users, user)
-	}
-	defer db.Close()
-
+	users := []model.UserInfo{}
+	// dbからユーザー一覧を取得
+	model.FindAllUser(&users)
 	output, _ := json.MarshalIndent(&users, "", "\t\t")
-	w.Header().Set("Access-Control-Allow-Origin", "http://localhost:3000")
-	w.Header().Set("Content-Type", "application/json")
 	w.Write(output)
 
 }
